+++ packages/plugin-workflow-approval/src/client/common/components/FormBlock.provider.tsx
@@ -15,7 +15,7 @@ import {
 import { createForm, RecursionField, useField, useFieldSchema } from '@tachybase/schema';
 
 import { useContextApprovalExecution } from '../../usage/common/contexts/approvalExecution';
-import { useContextApprovalRecords } from '../../usage/pc/block/todos-table/providers/ApprovalExecutions.provider';
+import { useContextApprovalRecords } from '../../usage/pc/block/table-todos/providers/ApprovalExecutions.provider';
 
 export const FormBlockProvider = (props) => {
   const { formType } = props;
