+++ packages/plugin-workflow-approval/src/client/usage/pc/block/table-initiations/CheckContent.schema.ts
