+++ packages/plugin-workflow-approval/src/client/usage/pc/common/process-columns/renderDetail.tsx
@@ -1,8 +1,8 @@
 import { useContext } from 'react';
 import { DatePicker, InputReadPretty, RecordProvider, useCurrentUserContext } from '@tachybase/client';
 
-import { ViewCheckLink as ViewCheckLinkInitiations } from '../../block/initiations-table/CheckLink.view';
-import { ViewCheckLink as ViewCheckLinkTodos } from '../../block/todos-table/CheckLink.view';
+import { ViewCheckLink as ViewCheckLinkInitiations } from '../../block/table-initiations/CheckLink.view';
+import { ViewCheckLink as ViewCheckLinkTodos } from '../../block/table-todos/CheckLink.view';
 import { ContextWithActionEnabled } from '../WithActionEnabled.provider';
 
 export function renderDetail(text, record, index) {
